Methods and variables are moved to separate files. First negative test is done
