+++ homework_27/test_data/variables.py
@@ -1,0 +1,5 @@
+url = "https://alexqa.netlify.app/.netlify"
+token = ("eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
+         ".eyJ1c2VySWQiOiIxMTY4NzI0NzY5MjY4MDM1NjM4MDIiLCJpYXQiOjE3MjM4MTgyNzYsImV4cCI6MTcyMzgyMTg3Nn0"
+         ".c7iXrW8NcyjlF4w4_MmQs5e2QFmPjR7wTHvj5OjE4ds")
+id_value = None
